refactor: experience-detail 도메인을 activities로 병합
