+++ src/domain/activities/components/activity/ImageGallery.tsx
