+++ src/domain/activities/components/activity/MobileActivityHeader.tsx
