+++ src/domain/activities/components/reservation/ParticipantCounter.tsx
